// libs/enums/cs.enum.ts
export enum CsStatus {
	ACTIVE = 'ACTIVE',
	DELETE = 'DELETE',
}

export enum CsType {
	NOTICE = 'NOTICE',
	FAQ = 'FAQ',
	INQUIRY = 'INQUIRY',
}

export enum CsCategory {
	PROPERTY = 'PROPERTY',
	PAYMENT = 'PAYMENT',
	BUYERS = 'BUYERS',
	AGENTS = 'AGENTS',
	MEMBERSHIP = 'MEMBERSHIP',
	COMMUNITY = 'COMMUNITY',
	OTHER = 'OTHER',
}

export enum InquiryStatus {
	PENDING = 'PENDING',
	ANSWERED = 'ANSWERED',
	CLOSED = 'CLOSED',
}